Made library compatible with both Unix and Windows versions of gpg-agent
